[MOV] mail_client_extension: split authentication and main controller code
This commit splits the code located in the mail_client_extension module into
two separate files, an "authentication.py" file containing authentication
related code and a "mail_client_extension.py" file containing everything else.

Purpose is to better organize the code and to avoid having a bloated main
controller.

This commit does not change anything functionally, it only moves and refactors
code.

Task-2382870

UPG-PR: https://github.com/odoo/upgrade/pull/2156
COM-PR: https://github.com/odoo/odoo/pull/66139
