+++ addons/mail_client_extension/controllers/__init__.py
@@ -1,2 +1,5 @@
-# -*- coding: utf-8 -*
-from . import main
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import authenticate
+from . import mail_plugin
